requirements.txt
Babel==0.9.6
Flask==0.10.1
Flask-Babel==0.8
Flask-Login==0.2.6
Flask-Mail==0.9.0
Flask-OpenID==1.1.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.8.4
Flask-WhooshAlchemy==0.55a
Jinja2==2.7
MarkupSafe==0.18
SQLAlchemy==0.7.9
Tempita==0.5.1
WTForms==1.0.4
Werkzeug==0.9.2
Whoosh==2.5.1
argparse==1.2.1
blinker==1.3
decorator==3.4.0
distribute==0.6.24
## FIXME: could not find svn URL in dependency_links for this package:
flup==1.0.3.dev-20110405
itsdangerous==0.22
python-openid==2.2.5
pytz==2013b
speaklater==1.3
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
